+++ lib/modules/craft/craft_detail_page.dart
@@ -192,16 +192,10 @@ class CraftDetailBasePage extends StatelessWidget {
                   title: Center(child: Text(S.of(context).view_illustration)),
                   contentPadding: EdgeInsets.zero,
                   onTap: () {
-                    Navigator.of(context).push(
-                      PageRouteBuilder(
-                        opaque: false,
-                        fullscreenDialog: true,
-                        pageBuilder: (context, _, __) => FullScreenImageSlider(
-                          imgUrls: [ce.illustration],
-                          placeholder: placeholder,
-                          allowSave: true,
-                        ),
-                      ),
+                    FullscreenImageViewer.show(
+                      context: context,
+                      urls: [ce.illustration],
+                      placeholder: placeholder,
                     );
                   },
                 ),